Extract sats-connect wallet methods from xverse wallet
