+++ packages/sdk/src/browser-wallets/sats-connect/constants.ts
