+++ packages/sdk/src/browser-wallets/sats-connect/utils.ts
